# Conflicts:
#	src/main/java/com/example/groupbuyingweb/GroupbuyingwebApplication.java
#	src/main/java/com/example/groupbuyingweb/domain/dto/request/GroupBuyingParticipationRequest.java

+++ src/main/java/com/example/groupbuyingweb/domain/entity/Member.java
@@ -36,7 +36,7 @@ public class Member {
     private Double longitude;
 
     @Column(nullable = false)
-    private Integer point;
+    private Double point;
     @Column(nullable = false)
     private Integer acornExp;
